# backend/openapi.json
{
  "components": {
    "schemas": {
      "AssignStrategyRequest": {
        "description": "POST /api/sessions/{id}/strategy 的请求体（保存新策略时改挂草稿会话）。",
        "properties": {
          "strategy_id": {
            "description": "新归属（策略 id）",
            "minLength": 1,
            "title": "Strategy Id",
            "type": "string"
          }
        },
        "required": [
          "strategy_id"
        ],
        "title": "AssignStrategyRequest",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "错误响应的统一形状——所有域错误都长这样（`{\"detail\": \"一句话\"}`）。\n\n各端点在 `responses=` 里引用它，把错误路径写进 OpenAPI 契约：\n前端生成类型时才能覆盖错误体，而不是只看到成功路径。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryMessageView": {
        "description": "一条历史消息。",
        "properties": {
          "attachment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "本轮整轮耗时毫秒（恢复历史后仍可显示；缺省 null）",
            "title": "Elapsed Ms"
          },
          "partial": {
            "default": false,
            "description": "True = 断流 / 报错时落盘的半截回复（界面标注「未完成」）",
            "title": "Partial",
            "type": "boolean"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "助手消息的思考过程全文（流式打标落盘；无思考为 null）",
            "title": "Reasoning"
          },
          "reasoning_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "本轮思考耗时毫秒（缺省 null）",
            "title": "Reasoning Ms"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "text",
          "attachment"
        ],
        "title": "HistoryMessageView",
        "type": "object"
      },
...
      "LabelRequest": {
        "description": "POST /api/label 的请求体。",
        "properties": {
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "图片（data URL 或纯 base64）；缺省纯文本轮",
            "title": "Image Base64"
          },
          "image_name": {
            "default": "image.png",
            "description": "图片原始文件名",
            "title": "Image Name",
            "type": "string"
          },
          "instruction": {
            "default": "",
            "description": "打标指令（纯图轮可空）",
            "title": "Instruction",
            "type": "string"
          },
          "prompt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "基础提示词 ID；续接时缺省沿用会话设置",
            "title": "Prompt Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "续接的会话 id；缺省新建",
            "title": "Session Id"
          },
          "skill_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "启用的 skill ID 清单；缺省沿用会话设置",
            "title": "Skill Ids"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "会话归属（策略 id 或 __new__ 草稿桶）：新建会话时盖章，续接时忽略（归属跟随既有会话）",
            "title": "Strategy Id"
          },
          "video_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "视频（data URL 或纯 base64）；与图片互斥（一期单素材/次）",
            "title": "Video Base64"
          },
          "video_fps": {
            "default": 2.0,
            "description": "视频抽帧 fps（整数值，1–10；端点对浮点 fps 判非法，整性在模型校验器兜底）",
            "maximum": 10.0,
            "minimum": 0.1,
            "title": "Video Fps",
            "type": "number"
          },
          "video_max_frames": {
            "default": 16,
            "description": "视频抽帧帧数上限",
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "Video Max Frames",
            "type": "integer"
          },
          "video_name": {
            "default": "video.mp4",
            "description": "视频原始文件名",
            "title": "Video Name",
            "type": "string"
          }
        },
        "title": "LabelRequest",
        "type": "object"
      },
      "LabelResponse": {
        "description": "POST /api/label 的响应体。",
        "properties": {
          "caption": {
            "title": "Caption",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "caption"
        ],
        "title": "LabelResponse",
        "type": "object"
      },
...
      "SessionSnapshotResponse": {
        "description": "GET /api/sessions/* 的响应体（恢复会话的完整快照）。",
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/HistoryMessageView"
            },
            "title": "Messages",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsView"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "会话归属（策略 id / __new__ 草稿桶）；无归属为 null",
            "title": "Strategy Id"
          }
        },
        "required": [
          "session_id",
          "settings",
          "messages"
        ],
        "title": "SessionSnapshotResponse",
        "type": "object"
      },
      "SettingsView": {
        "description": "会话当前设置（引用存各资产的稳定 ID）。",
        "properties": {
          "prompt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Id"
          },
          "skill_ids": {
            "items": {
              "type": "string"
            },
            "title": "Skill Ids",
            "type": "array"
          }
        },
        "required": [
          "prompt_id",
          "skill_ids"
        ],
        "title": "SettingsView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "AI 打标工具（发图 + 指令产出 caption，支持迭代改写）",
    "title": "Dataset Factory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/label": {
      "post": {
        "description": "跑一轮打标（带 session_id 即续接迭代改写）。",
        "operationId": "label_api_label_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "输入不合法（图片 / 视频不合法或互斥、提示词未选、空轮、端点配置缺失）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话或提示词不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话落盘等内部错误"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "模型端点调用失败"
          }
        },
        "summary": "Label",
        "tags": [
          "打标与会话"
        ]
      }
    },
    "/api/label/stream": {
      "post": {
        "description": "流式打标（SSE）：事件 = start → delta(kind=reasoning|content)… → done；失败发 error。",
        "operationId": "label_stream_api_label_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "输入不合法（同 /api/label；预备段失败走正常状态码，流中失败发 error 事件）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话或提示词不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Label Stream",
        "tags": [
          "打标与会话"
        ]
      }
    },
...
    "/api/sessions/latest": {
      "get": {
        "description": "最新会话快照（重启恢复入口）；一个会话都没有时 404。\n\n带 ``strategy_id`` 查询时按归属桶取最新（三期 v3：每策略各自的最近会话），\n该桶为空同样 404；不带时为全局最新（存量认领垫层用）。",
        "operationId": "latest_session_api_sessions_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "strategy_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "还没有任何会话"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Session",
        "tags": [
          "打标与会话"
        ]
      }
    },
    "/api/sessions/{session_id}": {
      "delete": {
        "description": "删除一个会话（事件流 + 附件 + 归属；有轮次在写时拒绝）。",
        "operationId": "delete_session_endpoint_api_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话有进行中的打标轮次"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session Endpoint",
        "tags": [
          "打标与会话"
        ]
      },
      "get": {
        "description": "某会话快照（设置 + 对话历史 + 归属）。",
        "operationId": "get_session_api_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "打标与会话"
        ]
      }
    },
    "/api/sessions/{session_id}/attachments/{name}": {
      "get": {
        "description": "取会话附件的文件字节（B5，2026-09-21 审计）：历史缩略图不再依赖内存 dataURL。\n\n安全口径与素材域的 /asset 同源：附件名经 sessions 域的单段安全名校验（路径穿越\n与非法字符在数据域拦下），只读、越界即 404。\n\nContent-Type 显式按扩展名给（PRD-0004）：``FileResponse`` 缺省靠 mimetypes 猜，\n猜不中的扩展名回落 octet-stream 会令 ``<video>``（历史封面 / 大图预览）拒播；\n映射与素材域 /asset 同一份（MIME 单一事实源在 llm.messages）。",
        "operationId": "session_attachment_api_sessions__session_id__attachments__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "附件名非法"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话或附件不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Attachment",
        "tags": [
          "打标与会话"
        ]
      }
    },
    "/api/sessions/{session_id}/strategy": {
      "post": {
        "description": "改挂会话归属（三期 v3）：保存新策略时把当前草稿会话从 ``__new__`` 挂到新 id。",
        "operationId": "assign_session_strategy_api_sessions__session_id__strategy_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "strategy_id 非法"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Session Strategy",
        "tags": [
          "打标与会话"
        ]
      }
    },
...
  }
}